+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **New features:**
 
 * st-form: Allow to display sections descriptions
+* st-foreground-notification: two way binding for visible input
 
 **Fixed bugs:**